review_api/requirements.txt
Automat==0.6.0
Flask==0.12.2
Flask-Login==0.4.0
Flask-RESTful==0.3.6
Jinja2==2.9.6
MarkupSafe==1.0
Twisted==17.9.0
Werkzeug==0.12.2
amqp==1.4.9
aniso8601==1.3.0
anyjson==0.3.3
argparse==1.2.1
attrs==17.2.0
billiard==3.3.0.23
boto==2.48.0
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
constantly==15.1.0
dateparser==0.7.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
kombu==3.0.37
lxml==4.1.0
pymongo==3.5.1
python-dateutil==2.6.1
pytz==2017.2
regex==2018.2.21
requests==2.18.4
six==1.11.0
tzlocal==1.5.1
uWSGI==2.0.15
urllib3==1.22
wsgiref==0.1.2
zope.interface==4.4.3
